Download command now checks hash of file
Download command first gets the size and hash from the server,
if the file doesn't exist on disk, its downloaded then hashed
to check if the file download was compromised. If it does exist
on disk already, then the file on disk is hashed to see if it
matches the server's hash. The user has the option to overwrite
the file if the hashes don't match.

Also added HELO command to connect_to_server so that the
server status shows my name.